Racing and Heuristic Changes
Support for mazebot's racing mode to test various heuristics, and draw
images after a race.

+++ MazeSolver/BreadthFirst.cs
@@ -6,7 +6,7 @@ namespace MazeSolver
 {
     class BreadthFirst : IHeuristic
     {
-        public int Score(Graph g, coordinate move)
+        public double Score(Graph g, coordinate move)
         {
             return 0;
         }
